requirements.txt
python-telegram-bot==21.7
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
pytz
